Rename SpecsHandler module to job_spec
After the I/O-orchestration refactor, this module holds a single
read-modify-write function (replace_tasks_in_job_spec) and no
SpecsHandler class, so the name was vestigial. Rename to job_spec.py
(and the test module to match): the module is the home for job-spec
operations, covering both reading the template spec and writing the
updated one. Manifest reading lives in Utils.read_dbt_manifest.
