+++ src/databricks_dbt_factory/main.py
@@ -5,7 +5,7 @@ from importlib import resources
 from pathlib import Path
 
 from databricks_dbt_factory.DbtFactory import DbtFactory
-from databricks_dbt_factory.SpecsHandler import replace_tasks_in_job_spec
+from databricks_dbt_factory.job_spec import replace_tasks_in_job_spec
 from databricks_dbt_factory.Utils import read_dbt_manifest
 from databricks_dbt_factory.DbtTask import DbtTaskOptions
 from databricks_dbt_factory.TaskFactory import (
